[DF-NONE] Fix warnings.
Former-commit-id: a2d79fba515c6fbd0eed29137e9472183ec704c7

+++ Sources/Internal/Render/RenderManagerGL.cpp
@@ -697,7 +697,7 @@ void RenderManager::DiscardFramebufferHW(uint32 attachments)
 void RenderManager::HWglDeleteBuffers(GLsizei count, const GLuint * buffers)
 {
     // TODO: this is, probably, temporary fix.
-    for(uint32 n = 0; n < count; ++n)
+    for(uint32 n = 0; n < (uint32)count; ++n)
     {
         if(bufferBindingId[0] == buffers[n])
         {
